Moving circuit breaker config to its own auto config
