+++ spring-cloud-netflix-core/src/main/java/org/springframework/cloud/netflix/hystrix/HystrixCircuitBreakerAutoConfiguration.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2013-2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.springframework.cloud.netflix.hystrix;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.cloud.client.circuitbreaker.CircuitBreakerBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import com.netflix.hystrix.Hystrix;
+
+/**
+ * @author Ryan Baxter
+ */
+@Configuration
+@ConditionalOnClass({ Hystrix.class })
+public class HystrixCircuitBreakerAutoConfiguration {
+	@Bean
+	@ConditionalOnMissingBean(HystrixCircuitBreakerConfigFactory.class)
+	public HystrixCircuitBreakerConfigFactory hystrixCircuitBreakerConfigFactory() {
+		return new HystrixCircuitBreakerConfigFactory.DefaultHystrixCircuitBreakerConfigFactory();
+	}
+
+	@Bean
+	@ConditionalOnMissingBean(CircuitBreakerBuilder.class)
+	public CircuitBreakerBuilder hystrixCircuitBreakerBuilder(HystrixCircuitBreakerConfigFactory hystrixCircuitBreakerConfigFactory) {
+		return new HystrixCircuitBreakerBuilder(hystrixCircuitBreakerConfigFactory);
+	}
+}
