+++ spring-cloud-netflix-core/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,7 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.cloud.netflix.hystrix.HystrixAutoConfiguration,\
-org.springframework.cloud.netflix.hystrix.security.HystrixSecurityAutoConfiguration
+org.springframework.cloud.netflix.hystrix.security.HystrixSecurityAutoConfiguration,\
+org.springframework.cloud.netflix.hystrix.HystrixCircuitBreakerAutoConfiguration
 
 org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker=\
 org.springframework.cloud.netflix.hystrix.HystrixCircuitBreakerConfiguration
